Move ChatPanel into component directory
