+++ apps/web/src/chat/ChatPanel/ChatPanel.stream-rendering.test.tsx
