+++ apps/web/src/chat/ChatPanel/index.ts
@@ -1,0 +1,1 @@
+export { ChatPanel } from "./ChatPanel";
